Add file structure for Set3 and add support for building tests.
Added the file and cmake structure to develop Set3 of challenges,
and also started generating a test executable to verify and ensure
backwards compatibility of the modifications that may or may not
be done to the already developed and used library routines.

Tests for challenges 1, 2, 3 and 4 have already been generated.